+++ internal/oplog/testing.go
@@ -2,26 +2,24 @@ package oplog
 
 import (
 	"crypto/rand"
-	"database/sql"
-	"fmt"
-	"strings"
 	"testing"
 
 	"github.com/golang-migrate/migrate/v4"
+	"github.com/hashicorp/boundary/internal/docker"
 	"github.com/hashicorp/boundary/internal/oplog/oplog_test"
 	wrapping "github.com/hashicorp/go-kms-wrapping"
 	"github.com/hashicorp/go-kms-wrapping/wrappers/aead"
 	"github.com/hashicorp/go-uuid"
 	"github.com/jinzhu/gorm"
-	"github.com/ory/dockertest/v3"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
 
-func testCleanup(t *testing.T, cleanupFunc func(), db *gorm.DB) {
+func testCleanup(t *testing.T, cleanupFunc func() error, db *gorm.DB) {
 	t.Helper()
-	cleanupFunc()
-	err := db.Close()
+	err := cleanupFunc()
+	assert.NoError(t, err)
+	err = db.Close()
 	assert.NoError(t, err)
 }
 
@@ -54,37 +52,14 @@ func testId(t *testing.T) string {
 	return id
 }
 
-// testInitDbInDocker initializes postgres within dockertest for the unit tests
-func testInitDbInDocker(t *testing.T) (cleanup func(), retURL string, err error) {
+func testInitDbInDocker(t *testing.T) (cleanup func() error, retURL string, err error) {
 	t.Helper()
-	pool, err := dockertest.NewPool("")
-	require.NoErrorf(t, err, "could not connect to docker: %w", err)
-
-	resource, err := pool.Run("postgres", "12", []string{"POSTGRES_PASSWORD=password", "POSTGRES_DB=boundary"})
-	require.NoErrorf(t, err, "could not start resource: %w", err)
-
-	c := func() {
-		err := testCleanupResource(t, pool, resource)
-		assert.NoError(t, err)
+	cleanup, retURL, _, err = docker.StartDbInDocker("postgres")
+	if err != nil {
+		t.Fatal(err)
 	}
-
-	url := fmt.Sprintf("postgres://postgres:password@localhost:%s?sslmode=disable", resource.GetPort("5432/tcp"))
-
-	err = pool.Retry(func() error {
-		db, err := sql.Open("postgres", url)
-		if err != nil {
-			return fmt.Errorf("error opening postgres dev container: %w", err)
-		}
-
-		if err := db.Ping(); err != nil {
-			return err
-		}
-		defer db.Close()
-		return nil
-	})
-	require.NoErrorf(t, err, "could not connect to docker: %w", err)
-	testInitStore(t, c, url)
-	return c, url, nil
+	testInitStore(t, cleanup, retURL)
+	return
 }
 
 // testWrapper initializes an AEAD wrapping.Wrapper for testing the oplog
@@ -101,30 +76,16 @@ func testWrapper(t *testing.T) wrapping.Wrapper {
 }
 
 // testInitStore will execute the migrations needed to initialize the store for tests
-func testInitStore(t *testing.T, cleanup func(), url string) {
+func testInitStore(t *testing.T, cleanup func() error, url string) {
 	t.Helper()
 	// run migrations
 	m, err := migrate.New("file://../db/migrations/postgres", url)
 	require.NoError(t, err, "Error creating migrations")
 
 	if err := m.Up(); err != nil && err != migrate.ErrNoChange {
-		cleanup()
+		if err := cleanup(); err != nil {
+			t.Fatalf("error cleaning up after migration failure: %v", err)
+		}
 		require.NoError(t, err, "Error running migrations")
 	}
-}
-
-// testCleanupResource will clean up the dockertest resources (postgres)
-func testCleanupResource(t *testing.T, pool *dockertest.Pool, resource *dockertest.Resource) error {
-	t.Helper()
-	var err error
-	for i := 0; i < 10; i++ {
-		err = pool.Purge(resource)
-		if err == nil {
-			return nil
-		}
-	}
-	if strings.Contains(err.Error(), "No such container") {
-		return nil
-	}
-	return fmt.Errorf("Failed to cleanup local container: %s", err)
 }
